Consolidated async and sync repositories
