--- src/DotNetToolkit.Repository/Factories/RepositoryFactoryAsync.cs
@@ -1,167 +1,0 @@
-﻿namespace DotNetToolkit.Repository.Factories
-{
-    using Interceptors;
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-
-    /// <summary>
-    /// An implementation of <see cref="IRepositoryFactoryAsync" />.
-    /// </summary>
-    /// <seealso cref="DotNetToolkit.Repository.Factories.IRepositoryFactoryAsync" />
-    public class RepositoryFactoryAsync : IRepositoryFactoryAsync
-    {
-        #region Fields
-
-        private readonly Func<IRepositoryContextAsync> _contextFactory;
-        private readonly IEnumerable<IRepositoryInterceptor> _interceptors;
-
-        #endregion
-
-        #region Constructors
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RepositoryFactoryAsync" /> class.
-        /// </summary>
-        /// <param name="contextFactory">The repository context.</param>
-        public RepositoryFactoryAsync(Func<IRepositoryContextAsync> contextFactory) : this(contextFactory, (IEnumerable<IRepositoryInterceptor>)null) { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RepositoryFactoryAsync" /> class.
-        /// </summary>
-        /// <param name="contextFactory">The repository context.</param>
-        /// <param name="interceptor">The interceptor.</param>
-        public RepositoryFactoryAsync(Func<IRepositoryContextAsync> contextFactory, IRepositoryInterceptor interceptor) : this(contextFactory, new List<IRepositoryInterceptor> { interceptor }) { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RepositoryFactoryAsync" /> class.
-        /// </summary>
-        /// <param name="contextFactory">The repository context.</param>
-        /// <param name="interceptors">The interceptors.</param>
-        public RepositoryFactoryAsync(Func<IRepositoryContextAsync> contextFactory, IEnumerable<IRepositoryInterceptor> interceptors)
-        {
-            if (contextFactory == null)
-                throw new ArgumentNullException(nameof(contextFactory));
-
-            _contextFactory = contextFactory;
-            _interceptors = interceptors ?? Enumerable.Empty<IRepositoryInterceptor>();
-        }
-
-        #endregion
-
-        #region Implementation of IRepositoryFactory
-
-        /// <summary>
-        /// Creates a new repository for the specified entity type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <returns>The new repository.</returns>
-        public IRepository<TEntity> Create<TEntity>() where TEntity : class
-        {
-            return CreateAsync<TEntity>();
-        }
-
-        /// <summary>
-        /// Creates a new repository for the specified entity and primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey">The type of the key primary key value.</typeparam>
-        /// <returns>The new repository.</returns>
-        public IRepository<TEntity, TKey> Create<TEntity, TKey>() where TEntity : class
-        {
-            return CreateAsync<TEntity, TKey>();
-        }
-
-        /// <summary>
-        /// Creates a new repository for the specified entity and a composite primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey1">The type of the first part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey2">The type of the second part of the composite primary key.</typeparam>
-        /// <returns>The new repository.</returns>
-        public IRepository<TEntity, TKey1, TKey2> Create<TEntity, TKey1, TKey2>() where TEntity : class
-        {
-            return CreateAsync<TEntity, TKey1, TKey2>();
-        }
-
-        /// <summary>
-        /// Creates a new repository for the specified entity and a composite primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey1">The type of the first part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey2">The type of the second part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey3">The type of the third part of the composite primary key.</typeparam>
-        /// <returns>The new repository.</returns>
-        public IRepository<TEntity, TKey1, TKey2, TKey3> Create<TEntity, TKey1, TKey2, TKey3>() where TEntity : class
-        {
-            return CreateAsync<TEntity, TKey1, TKey2, TKey3>();
-        }
-
-        /// <summary>
-        /// Creates a new repository for the specified type.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns>The new repository.</returns>
-        public T CreateInstance<T>() where T : class
-        {
-            var args = new List<object> { _contextFactory() };
-
-            if (_interceptors.Any())
-                args.Add(_interceptors);
-
-            return (T)Activator.CreateInstance(typeof(T), args.ToArray());
-        }
-
-        #endregion
-
-        #region Implementation of IRepositoryFactoryAsync
-
-        /// <summary>
-        /// Creates a new asynchronous repository for the specified entity type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <returns>The new asynchronous repository.</returns>
-        public IRepositoryAsync<TEntity> CreateAsync<TEntity>() where TEntity : class
-        {
-            return CreateInstance<RepositoryAsync<TEntity>>();
-        }
-
-        /// <summary>
-        /// Creates a new asynchronous repository for the specified entity and primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey">The type of the key primary key value.</typeparam>
-        /// <returns>The new asynchronous repository.</returns>
-        public IRepositoryAsync<TEntity, TKey> CreateAsync<TEntity, TKey>() where TEntity : class
-        {
-            return CreateInstance<RepositoryAsync<TEntity, TKey>>();
-        }
-
-        /// <summary>
-        /// Creates a new asynchronous repository for the specified entity and a composite primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey1">The type of the first part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey2">The type of the second part of the composite primary key.</typeparam>
-        /// <returns>The new asynchronous repository.</returns>
-        public IRepositoryAsync<TEntity, TKey1, TKey2> CreateAsync<TEntity, TKey1, TKey2>() where TEntity : class
-        {
-            return CreateInstance<RepositoryAsync<TEntity, TKey1, TKey2>>();
-        }
-
-        /// <summary>
-        /// Creates a new asynchronous repository for the specified entity and a composite primary key type.
-        /// </summary>
-        /// <typeparam name="TEntity">The type of the entity.</typeparam>
-        /// <typeparam name="TKey1">The type of the first part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey2">The type of the second part of the composite primary key.</typeparam>
-        /// <typeparam name="TKey3">The type of the third part of the composite primary key.</typeparam>
-        /// <returns>The new asynchronous repository.</returns>
-        public IRepositoryAsync<TEntity, TKey1, TKey2, TKey3> CreateAsync<TEntity, TKey1, TKey2, TKey3>() where TEntity : class
-        {
-            return CreateInstance<RepositoryAsync<TEntity, TKey1, TKey2, TKey3>>();
-        }
-
-        #endregion
-    }
-}
